Put auth repository api from its contract in different files
